proof of concept for pod

+++ modules/display_parallel/firmware/dspic_ntg128641a/common.h
@@ -28,8 +28,8 @@ void __delay32(u32 cycles);
 #define delay_us(us)            __delay32((u32)(us * 60))
 #define delay_ms(ms)            __delay32((u32)(ms * 59881))
 
-int mpow(int x, int exp);
-char *mprint_int(char *s, int n, int base, int digits);
-char *mprint_float(char *str, float val, int pre_digits, int post_digits);
+s16 mpow(s16 x, s16 exp);
+char *mprint_int(char *s, s16 n, s16 base, s16 digits);
+char *mprint_float(char *str, float val, s16 pre_digits, s16 post_digits);
 
 #endif
